force reload serverList on first fetchServers
